chore(domain): update CONTEXT.md with knowledge graph terminology and refactor extraction types
